Created app for offers and Offer model, serializer etc. replaced there.
