+++ backend/internet_auction/urls.py
@@ -9,6 +9,7 @@ urlpatterns = [
     path('api/', include('auction.urls')),
     path('api/', include('item.urls')),
     path('api/', include('lot.urls')),
+    path('api/', include('offer.urls')),
     path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
     path('api/schema/swagger-ui/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
     path('api/schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
