+++ backend/lot/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import Lot, Offer
+from .models import Lot
 
 admin.site.register(Lot)
-admin.site.register(Offer)
