+++ backend/offer/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Offer
+
+
+class OfferSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Offer
+        fields = '__all__'
